+++ History.md
@@ -1,5 +1,8 @@
 # History
 
+#### 2021-01-22
+* Add Exception with ResponseStatus annotation and use it in Controllers. Closes #43.
+
 #### 2021-01-20
 * Add ability to add or update Answer. Closes #41.
 * Fix updating Question. Closes #42.